Use new way to encode features the BSGL algorithms
In learner, there is a module to map the state observation to features.
A variety types of features are supported.  Previously, the feature are
decoded based on their aboslution position, which is error-prone.
Commit 32e2ea2ab20f857476eefc26585dd9c374162e9e introduced a better way
to encode and decode feature. This commit updates the bsgl algorithms to
use the new approach.

TESTED=blaze test. sample run ./blaze run examples/garnet/garnetexample.py