Fix the bsgl example for TAC simulation
The previous setup of bsgl for tac simulation was wrong.

BSGL methods used a different type of feature: the state feature was
generated randomly. The state-action feature was generated by padding
zeros to the state feature. For other methods (TD, LSTD, HESSIAN), the
state action feature is usually defined as expected improvement of
applying an action at a state.

As a result, bsgl methods need to use a different type of task and uses
PolicyFeatureValue module. The type of task and module need to be
matched.