Update maze avg reward examples for tac results
The previous avg reward example was a very challenging problem. There
are much more empty states then goal states. Since the probably to reach
goal is too low, the goal reward has no influence to the result.

In this update
1. There are still some rectangular regions. Some of them are unsafe
regions, some of them are goal regions. The reward to goal region is 1
and unsafe region is -1
2. The observed feature is a normalized feature
3. Change the sessionSize to be 100, and update threshold for hessian
algorithm.